pyproject.toml
[build-system]
requires = [ "setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "workspace-mcp"
version = "1.18.0"
description = "Comprehensive, highly performant Google Workspace Streamable HTTP & SSE MCP Server for Calendar, Gmail, Docs, Sheets, Slides & Drive"
readme = "README.md"
keywords = [ "mcp", "google", "workspace", "llm", "ai", "claude", "model", "context", "protocol", "server"]
license = "MIT"
requires-python = ">=3.10"
dependencies = [
 "fastapi>=0.115.12",
 "fastmcp>=3.2.0",
 "google-api-python-client>=2.168.0",
 "google-auth-httplib2>=0.2.0",
 "google-auth-oauthlib>=1.2.2",
 "httpx>=0.28.1",
 "py-key-value-aio>=0.3.0",
 "pyjwt>=2.12.0",
 "python-dotenv>=1.1.0",
 "pyyaml>=6.0.2",
 "cryptography>=45.0.0",
 "defusedxml>=0.7.1",
 "pytz>=2026.1.post1",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: Communications :: Chat",
    "Topic :: Office/Business",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed"
]

[[project.authors]]
name = "Taylor Wilsdon"
email = "taylor@taylorwilsdon.com"

[project.urls]
Homepage = "https://workspacemcp.com"
Repository = "https://github.com/taylorwilsdon/google_workspace_mcp"
Documentation = "https://github.com/taylorwilsdon/google_workspace_mcp#readme"
Issues = "https://github.com/taylorwilsdon/google_workspace_mcp/issues"
Changelog = "https://github.com/taylorwilsdon/google_workspace_mcp/releases"

[project.scripts]
workspace-mcp = "main:main"
workspace-cli = "core.cli:main"

[project.optional-dependencies]
disk = [
    "py-key-value-aio[filetree]>=0.3.0",
]
valkey = [
    "py-key-value-aio[valkey]>=0.3.0",
]
test = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.23.0",
    "requests>=2.32.3",
]
release = [
    "tomlkit>=0.13.3",
    "twine>=5.0.0",
]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.23.0",
    "requests>=2.32.3",
    "ruff>=0.12.4",
    "tomlkit>=0.13.3",
    "twine>=5.0.0",
]

[dependency-groups]
disk = [
    "py-key-value-aio[filetree]>=0.3.0",
]
valkey = [
    "py-key-value-aio[valkey]>=0.3.0",
]
test = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.23.0",
    "requests>=2.32.3",
]
release = [
    "tomlkit>=0.13.3",
    "twine>=5.0.0",
]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.23.0",
    "requests>=2.32.3",
    "ruff>=0.12.4",
    "tomlkit>=0.13.3",
    "twine>=5.0.0",
]

[tool.setuptools]
py-modules = [ "fastmcp_server", "main"]

[tool.setuptools.packages.find]
where = ["."]
exclude = ["tests*", "docs*", "build", "dist"]

[tool.pytest.ini_options]
addopts = "--ignore=tests/gappsscript/manual_test.py"

[tool.setuptools.package-data]
core = ["tool_tiers.yaml"]
